Add files via upload
The project focuses on building a recommendation engine for Netflix, using user ratings and preferences to suggest movies. It analyzes a dataset containing customer IDs, movie names, genres, and ratings to identify popular genres and determine which ones received the best and worst reviews. A personalized model is created to recommend the most suitable movies for each user based on their viewing history and interests. The overall goal is to improve user engagement by providing tailored movie suggestions.